+++ app/src/main/res/layout/activity_signup.xml
@@ -4,7 +4,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:paddingHorizontal="27dp">
+        android:paddingHorizontal="20dp">
     <TextView
             android:id="@+id/tome_logo"
             android:layout_width="wrap_content"
@@ -16,7 +16,6 @@
             android:layout_marginTop="27dp"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintStart_toStartOf="parent"/>
-
     <TextView
             android:id="@+id/tv_signup"
             android:layout_width="wrap_content"
@@ -45,6 +44,73 @@
             android:layout_marginTop="27dp"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintStart_toStartOf="parent"/>
+    <TextView
+            android:id="@+id/tv_nickname"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/nickname"
+            android:textAppearance="@style/text_body2"
+            android:textColor="@color/color_disabled1"
+            android:layout_marginTop="90dp"
+            app:layout_constraintTop_toBottomOf="@id/tv_signup"
+            app:layout_constraintStart_toStartOf="parent"/>
+    <EditText
+            android:id="@+id/et_nickname"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toBottomOf="@id/tv_nickname"
+            app:layout_constraintBottom_toTopOf="@id/border_nickname"
+            android:background="@color/color_transparent"/>
+    <View
+            android:id="@+id/border_nickname"
+            android:layout_width="match_parent"
+            android:layout_height="1dp"
+            android:background="@color/color_disabled1"
+            android:layout_marginTop="39dp"
+            app:layout_constraintTop_toBottomOf="@id/tv_nickname"/>
+    <TextView
+            android:id="@+id/tv_birthdate"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/birthdate"
+            android:textAppearance="@style/text_body2"
+            android:textColor="@color/color_disabled1"
+            android:layout_marginTop="55dp"
+            app:layout_constraintTop_toBottomOf="@id/tv_nickname"
+            app:layout_constraintStart_toStartOf="parent"/>
+    <EditText
+            android:id="@+id/et_birthdate"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            app:layout_constraintTop_toBottomOf="@id/tv_birthdate"
+            app:layout_constraintBottom_toTopOf="@id/border_birthdate"
+            android:background="@color/color_transparent"/>
+    <View
+            android:id="@+id/border_birthdate"
+            android:layout_width="match_parent"
+            android:layout_height="1dp"
+            android:background="@color/color_disabled1"
+            android:layout_marginTop="39dp"
+            app:layout_constraintTop_toBottomOf="@id/tv_birthdate"/>
+    <CheckBox
+            android:id="@+id/chk_policy"
+            android:layout_width="16dp"
+            android:layout_height="16dp"
+            android:background="@color/color_transparent"
+            app:layout_constraintTop_toTopOf="@id/tv_policy"
+            app:layout_constraintBottom_toBottomOf="@id/tv_policy"
+            app:layout_constraintStart_toStartOf="parent"/>
+    <TextView
+            android:id="@+id/tv_policy"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/signup_policy"
+            android:textAppearance="@style/text_body2"
+            android:textColor="@color/color_font3"
+            android:layout_marginTop="55dp"
+            android:layout_marginStart="8dp"
+            app:layout_constraintStart_toEndOf="@id/chk_policy"
+            app:layout_constraintTop_toBottomOf="@id/tv_birthdate"/>
     <Button
             android:id="@+id/btn_sign_up"
             android:layout_width="match_parent"
@@ -55,9 +121,8 @@
             android:textAppearance="@style/text_body1"
             android:textColor="@color/color_font3"
             android:layout_marginBottom="46dp"
+            android:layout_marginHorizontal="6dp"
             app:layout_constraintBottom_toBottomOf="parent"
-            android:enabled="false"
+            android:enabled="true"
             style="@style/Widget.AppCompat.Button.Borderless"/>
-
-
 </androidx.constraintlayout.widget.ConstraintLayout>
